Refactor database settings UI
