+++ src/components/DatabaseStatusPanel.tsx
@@ -41,7 +41,6 @@ export default function DatabaseStatusPanel() {
     <div className="bg-white rounded-lg shadow-md p-6 border border-gray-200 space-y-4">
       <div className="flex items-center space-x-3">
         <DatabaseIcon className="h-6 w-6" style={{ color: '#F29400' }} />
-        <h3 className="text-lg font-medium text-gray-900">Datenbankstatus</h3>
       </div>
       <div className="flex items-center space-x-2">
         {renderIcon()}
